:whale: :recycle: add support for aws-cliv2 + awslocal bin
The binary needs to use the venv python installed to avoid
using the one from the system. A new version must remove
this step to avoid confusion and end up having 2 python
installations in the same docker image.

For now, since there are multiple python instances, the
system is pointing to an older version, while venv has
a newer version python 3.13. So, the awslocal-docker
has the proper call as the order of PATH needs to be
respected so that the AWS-cliv2 is loaded beore aws-v1
added by the python bindings.